+++ src/main/java/calculator/view/input/MenuInput.java
@@ -7,11 +7,11 @@ import static calculator.view.util.regex.RegexInputUtil.checkWrong;
 
 public class MenuInput implements BaseInput {
 
-    public Long askMenuId() {
+    public int askMenuId() {
         String menuId = read();
         if (checkWrong(menuId)) {
             throw new InputMismatchException(MENU_INPUT_NULL_EXCEPTION.getMessage());
         }
-        return Long.parseLong(menuId);
+        return Integer.parseInt(menuId);
     }
 }
